pyproject.toml
[project]
name = "dnd-cards"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "reportlab>=4.4.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
]

[project.scripts]
dnd-cards = "dnd_cards.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with -m 'not slow')"]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "TID252"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.mypy]
strict = true
mypy_path = "src"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "types-pyyaml>=6.0.12.20260408",
]
